extract the filter bar
